backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
greenlet==3.1.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
boto3==1.34.34
qrcode[pil]==7.4.2
vobject==0.9.9
email-validator==2.1.0
mangum==0.17.0
